+++ app/src/main/java/com/example/musiclk/Util/MD5Utils.java
@@ -1,4 +1,4 @@
-package com.example.musiclk;
+package com.example.musiclk.Util;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
