+++ app/src/main/java/com/example/musiclk/activity/MainActivity.java
@@ -1,8 +1,10 @@
-package com.example.musiclk;
+package com.example.musiclk.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.musiclk.R;
 
 public class MainActivity extends AppCompatActivity {
